+++ lib/cli_runner.dart
@@ -1,0 +1,1 @@
+export 'package:dart_code_metrics/src/cli/cli_runner.dart';