+++ lib/src/cli/commands/analyze_command.dart
@@ -7,8 +7,10 @@ import 'package:collection/collection.dart';
 import '../../analyzers/lint_analyzer/lint_analyzer.dart';
 import '../../analyzers/lint_analyzer/metrics/metrics_factory.dart';
 import '../../analyzers/lint_analyzer/metrics/models/metric_value_level.dart';
+import '../../analyzers/lint_analyzer/models/lint_file_report.dart';
 import '../../analyzers/lint_analyzer/models/severity.dart';
 import '../../analyzers/lint_analyzer/reporters/lint_report_params.dart';
+import '../../analyzers/lint_analyzer/reporters/reporters_list/json/lint_json_reporter.dart';
 import '../../analyzers/lint_analyzer/utils/report_utils.dart';
 import '../../config_builder/config_builder.dart';
 import '../../config_builder/models/deprecated_option.dart';
@@ -44,25 +46,28 @@ class AnalyzeCommand extends BaseCommand {
       ..isVerbose = isVerbose
       ..progress.start('Analyzing');
 
-    final parsedArgs = ParsedArguments(
-      excludePath: argResults[FlagNames.exclude] as String,
-      metricsConfig: {
-        for (final metric in getMetrics(config: {}))
-          if (argResults.wasParsed(metric.id))
-            metric.id: argResults[metric.id] as Object,
-      },
-    );
+    final parsedArgs = ParsedArguments.fromArgs(argResults);
 
     final config = ConfigBuilder.getLintConfigFromArgs(parsedArgs);
 
     final lintAnalyzerResult = await _analyzer.runCliAnalysis(
       argResults.rest,
-      argResults[FlagNames.rootFolder] as String,
+      parsedArgs.rootFolder,
       config,
       sdkPath: findSdkPath(),
     );
 
     _logger.progress.complete('Analysis is completed. Preparing the results:');
+
+    final jsonReportPath = parsedArgs.jsonReportPath;
+    if (jsonReportPath != null) {
+      final jsonReporter =
+          LintJsonReporter.toFile(jsonReportPath, parsedArgs.rootFolder);
+      await jsonReporter.report(
+        lintAnalyzerResult,
+        summary: _analyzer.getSummary(lintAnalyzerResult),
+      );
+    }
 
     await _analyzer
         .getReporter(
@@ -76,6 +81,10 @@ class AnalyzeCommand extends BaseCommand {
           additionalParams: LintReportParams(congratulate: !isNoCongratulate),
         );
 
+    _checkSeverity(lintAnalyzerResult);
+  }
+
+  void _checkSeverity(Iterable<LintFileReport> lintAnalyzerResult) {
     if (hasIssueWithSeverity(lintAnalyzerResult, Severity.error)) {
       exit(3);
     } else if ((argResults[FlagNames.fatalWarnings] as bool) &&
@@ -133,6 +142,12 @@ class AnalyzeCommand extends BaseCommand {
         help: 'Write HTML output to OUTPUT.',
         valueHelp: 'OUTPUT',
         defaultsTo: 'metrics',
+      )
+      ..addOption(
+        FlagNames.jsonReportPath,
+        help: 'Path to the JSON file with the output of the analysis.',
+        valueHelp: 'path/to/file.json',
+        defaultsTo: null,
       );
   }