+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 * feat: **Breaking change** rename `member-ordering-extended` to `member-ordering`, discarding the old implementation.
+* feat: support report to the json file option for the `analyze` command.
+* feat: make CliRunner a part of public API in order to support transitive executable calls use-case.
 * feat: add static code diagnostic [`avoid-cascade-after-if-null`](https://dartcodemetrics.dev/docs/rules/common/avoid-cascade-after-if-null).
 
 ## 4.21.2